// src/types/Users.tsx
export interface User {
    id: number;
    name: string;
    username: string;
    email: string;
    address: {
        street: string;
        suite: string;
        city: string;
        zipcode: string;
        geo: {
            lat: string;
            lng: string;
        };
    };
    phone: string;
    website: string;
    company: {
        name: string;
        catchPhrase: string;
        bs: string;
    };
};

export type UserResponse = User[];

export interface FormData {
    name: string;
    email: string;
    phone: string;
    address: string;
    website: string;
    company: string;
}